chore: 清理全部 Kotlin 编译警告
- BackupRepository: encodeToStream/decodeFromStream 加 @OptIn(ExperimentalSerializationApi)；settings 非空，删除恒真判空
- ScaleBleEngine: CCCD/档案写入迁移到 API 33+ 新重载（旧系统回落弃用写法）；connectGatt 新替代品仅 API 37+，定向 @Suppress 保留；去掉多余安全调用
- 四个 BottomSheet 迁移 rememberBottomSheetState(Hidden)（官方 ReplaceWith 映射）
- LocalLifecycleOwner 导入迁至 androidx.lifecycle.compose；WidgetRepository 去掉冗余 toInt()

+++ app/src/main/java/com/example/weight/data/widget/WidgetRepository.kt
@@ -45,7 +45,7 @@ class WidgetRepository(private val recordDao: RecordDao) {
         // 三次独立查询并行执行；Room suspend 自带 IO 调度，不再额外 withContext
         val currentDeferred = async { recordDao.getLastData() }
         val dailyDeferred = async {
-            recordDao.getDailyMinWeightSince(getStartTimeForLastDays(PREDICT_WINDOW_DAYS.toInt())).first()
+            recordDao.getDailyMinWeightSince(getStartTimeForLastDays(PREDICT_WINDOW_DAYS)).first()
         }
         val firstDeferred = async { recordDao.getFirstData() }
         val recordDaysDeferred = async { recordDao.getRecordDaysFlow().first() }
